#990 ♻ revert datakind raw (#1000)
* Readme

* revert raw from appdef

* revert raw from parser

* revert raw from istructsmem\dynobuffer

* revert raw from istructsmem

* revert raw from istructsmem
fix pkg\processor\command sys.Raw test

* fix pkg\processor\query sys.Raw test

+++ pkg/parser/impl_test.go
@@ -456,7 +456,7 @@ func Test_Varchar(t *testing.T) {
 	TYPE CDoc1 (
 		Oversize varchar(%d)
 	);
-	`, appdef.MaxFieldLength+1, appdef.MaxFieldLength+1))
+	`, uint32(appdef.MaxFieldLength)+1, uint32(appdef.MaxFieldLength)+1))
 	require.NoError(err)
 	pkg, err := BuildPackageSchema("pkg/test", []*FileSchemaAST{fs})
 	require.NoError(err)
